+++ lib/middleman-presentation/presentation_config.rb
@@ -49,6 +49,8 @@ module Middleman
       option :edit, false
       option :editor_command, process_environment.fetch('EDITOR', 'vim')
       option :error_on_duplicates, true
+
+      option :force_create_presentation, false
     end
   end
 end