+++ RELEASE_NOTES.md
@@ -2,6 +2,7 @@
   * F: Introduce Plugin-Support
   * F: Add helpers to show available "assets", "frontend components", "plugins", ...
   * F: Build command added
+  * I: Enforce flag for "create presentation"-command
   * I: Export command now creates a generated presentation as zip file
   * I: Move helpers to helper-gem
   * I: Localize cli